+++ libavcodec/libcodec2/fdmdv.c
@@ -86,7 +86,7 @@ struct FDMDV *fdmdv_create(int Nc) {
   f->rx_test_bits_mem = (int *)MALLOC(sizeof(int) * f->ntest_bits);
   assert(f->rx_test_bits_mem != NULL);
   for (i = 0; i < f->ntest_bits; i++) f->rx_test_bits_mem[i] = 0;
-  assert((sizeof(test_bits) / sizeof(int)) >= f->ntest_bits);
+  assert((sizeof(test_bits) / sizeof(int)) >= (long long unsigned int)f->ntest_bits);
 
   f->old_qpsk_mapping = 0;
